style: search 곡 제목 넘침 수정
